firebase connection and modes
added modes 0,1,2,3,4.

mode_set() is used to change modes, so the code to update the firebase is in one place.

reset_robot() for when mode is set to 1.

streams waiting and active task, and reacts accordingly. when waiting is set to 0 while it is on mode=3, it changes modes. when activeTask is changed, it puts those nodes to an "order" vector, from where it will be visiting places.

code to turn the FinalGraphOutput into a single string is written, also, the code to send the final graph to Database is written, but as is they are not called anywhere.

erased redundant modes.

added basic logic to know when it needs to go back to start node and when it needs to go to its next task(checks if order vector is empty)